Seth-added-function-3-to-main
I imported the functions fro "function3Final". Then wrote the code that if a user selects "1" or "choose team" it runs the functions to show a menu of teams and assign the selected ones to home team and away team.

I also modified the functions on function 1 + 2
intro function : rstrip so that it doesn't keep trace of spaces if one is entered at the end of the name

menu: int() just so that the choice doesn't come back as a string.